chore(backend): Improve paths for upload files object
