+++ packages/backend/src/core/files/helpers/upload/dto/delete.args.ts
@@ -1,13 +1,5 @@
-import { ArgsType, Field } from '@nestjs/graphql';
-
-@ArgsType()
-export class DeleteCoreFilesArgs {
-  @Field(() => String)
+export interface DeleteCoreFilesArgs {
   dir_folder: string;
-
-  @Field(() => String)
   file_name: string;
-
-  @Field(() => Boolean, { nullable: true })
   secure?: boolean;
 }
