+++ packages/backend/src/core/files/helpers/upload/dto/upload.args.ts
@@ -1,4 +1,6 @@
-import { FileUpload } from '@/graphql-upload';
+import { Field, InputType } from '@nestjs/graphql';
+
+import { FileUpload, GraphQLUpload } from '@/graphql-upload';
 
 export interface UploadCoreFilesArgs {
   acceptMimeType: string[];
@@ -7,4 +9,13 @@ export interface UploadCoreFilesArgs {
   maxUploadSizeBytes: number;
   plugin: string;
   secure?: boolean | null;
+}
+
+@InputType()
+export class UploadWithKeepCoreFilesArgs {
+  @Field(() => GraphQLUpload, { nullable: true })
+  file?: Promise<FileUpload>;
+
+  @Field(() => Boolean, { nullable: true })
+  keep?: boolean;
 }
